requirements.txt
discord.py
aiosqlite>=0.19.0
aiohttp>=3.9.0
python-dateutil>=2.8.2
psutil
flask